Add hotel to display, change int variables to double ones in hotelCost, apply same chagnes in header.h

+++ Travel_Expenses/hotelCost.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include "header.h"
 
-void hotelCost(int *total, int days){
-    int dailyHotelExpense = 90;
+void hotelCost(double *total, int days){
+    double dailyHotelExpense = 90.0;
     total[0] = days*dailyHotelExpense;
 
     printf("How much did you spend on lodging? ");
